Crate servo_webvtt

Structs

IncrementalWebVTTParser

WebVttCue

https://w3c.github.io/webvtt/#webvtt-cue

WebVttCueSize

https://w3c.github.io/webvtt/#webvtt-cue-size This is a struct, since the default value is 100

Enums

MostSignificantUnits 🔒

ParserState 🔒

WebVttLineAlignment

https://w3c.github.io/webvtt/#webvtt-cue-line-alignment

WebVttLineAndPositionSetting

https://w3c.github.io/webvtt/#webvtt-cue-position

WebVttParserError

WebVttPositionAlignment

https://w3c.github.io/webvtt/#webvtt-cue-position-alignment

WebVttSnapToLines

https://w3c.github.io/webvtt/#webvtt-cue-snap-to-lines-flag This is an enum, since the default value is true

WebVttTextAlignment

https://w3c.github.io/webvtt/#webvtt-cue-text-alignment

WebVttWritingDirection

https://w3c.github.io/webvtt/#webvtt-cue-writing-direction

Statics

WEB_VTT_PERCENTAGE_GRAMMAR 🔒

https://w3c.github.io/webvtt/#webvtt-percentage

Traits

WebVttParserSink

Functions

collect_ascii_digits 🔒

collect_for_closure 🔒

collect_webvtt_cue_timings_and_settings 🔒

https://w3c.github.io/webvtt/#collect-webvtt-cue-timings-and-settings

collect_webvtt_timestamp 🔒

https://w3c.github.io/webvtt/#collect-a-webvtt-timestamp

parse_a_percentage_string 🔒

https://w3c.github.io/webvtt/#parse-a-percentage-string

parse_the_webvtt_cue_settings 🔒

https://w3c.github.io/webvtt/#parse-the-webvtt-cue-settings

skip_whitespace 🔒

Type Aliases

ParserUpdate