servo_webvtt::cue

Module settings

Structs

WebVttCue

https://w3c.github.io/webvtt/#webvtt-cue

WebVttCueSize

https://w3c.github.io/webvtt/#webvtt-cue-size This is a struct, since the default value is 100

Enums

WebVttLineAlignment

https://w3c.github.io/webvtt/#webvtt-cue-line-alignment

WebVttLineAndPositionSetting

https://w3c.github.io/webvtt/#webvtt-cue-position

WebVttPositionAlignment

https://w3c.github.io/webvtt/#webvtt-cue-position-alignment

WebVttSnapToLines

https://w3c.github.io/webvtt/#webvtt-cue-snap-to-lines-flag This is an enum, since the default value is true

WebVttTextAlignment

https://w3c.github.io/webvtt/#webvtt-cue-text-alignment

WebVttWritingDirection

https://w3c.github.io/webvtt/#webvtt-cue-writing-direction